refactor: formatCompact/formatChange utils 중앙화 — 로컬 중복 함수 제거
- format.ts에 formatCompact, formatChange 추가
- UnifiedSummaryCards: 로컬 formatAmount/formatLargeAmount → utils import
- MonthlyComparison: 로컬 formatCompact/formatChange → utils import
- HeroSummary: formatWon 내부를 formatCompact로 위임 (₩0 엣지케이스 래퍼 유지)
